FirebaseAuth エラーコード定義の分離と配置の適正化
